UserController OK

+++ src/controllers/UserController.js
@@ -6,8 +6,14 @@ const userSchema = yup.object().shape({
   name: yup.string().required("O nome e obrigatorio"),
   email: yup.string().email().required("O email e obrigatorio"),
   password: yup.string().required("A senha é obrigatoria"),
-  // birth_date: yup.date().required("A data de nascimento e obrigatoria"),
-  // user_type:
+  birth_date: yup.string().required("A data de nascimento e obrigatoria"),
+  user_type: yup
+    .string()
+    .oneOf(
+      ["guia", "turista"],
+      "O tipo de usuário deve ser 'guia' ou 'turista'"
+    )
+    .required("O tipo de usuário é obrigatório"),
 });
 
 class UserController {
@@ -27,6 +33,13 @@ class UserController {
       });
 
       const { name, email, password, birth_date, user_type } = req.body;
+
+      if (!birth_date.match(/\d{4}-\d{2}-\d{2}/gm)) {
+        return res.status(400).json({
+          message: "Formato correto da data de nascimento e ANO-MES-DIA",
+        });
+      }
+
       const existingUser = await User.findOne({
         where: { email },
       });
@@ -48,12 +61,14 @@ class UserController {
       res.status(201).json({ message: "Usuario criado com sucesso", user });
     } catch (error) {
       console.log(error.message);
-      res.status(500).json(error.message);
+      res.status(500).json({ error: "Não possível e cadastrar o usuario" });
     }
   }
   async delete(req, res) {
     try {
       const { id } = req.params;
+      const authenticatedUserId = req.payload.sub;
+
       const user = await User.findByPk(id);
       const tour = await Tour.findAll({
         where: { user_id: id },
@@ -61,17 +76,25 @@ class UserController {
       if (!user) {
         return res.status(404).json({ message: "Usuario não encontrado" });
       }
+
       if (tour.length > 0) {
         return res.status(403).json({
-          message: "Usuário não se pode eliminar, tem locales asociados!",
-        });
-      } else {
-        await User.destroy({
-          where: {
-            id,
-          },
+          message:
+            "Usuário não pode ser eliminado, pois tem locais associados!",
         });
       }
+
+      if (user.id !== authenticatedUserId) {
+        return res.status(403).json({
+          message: "Você não tem permissão para eliminar este usuário.",
+        });
+      }
+
+      await User.destroy({
+        where: {
+          id,
+        },
+      });
 
       res.status(200).json({ message: "Usuario eliminado com sucesso" });
     } catch (error) {
